% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.mapping-package.R
\docType{package}
\name{gsm.mapping-package}
\alias{gsm.mapping}
\alias{gsm.mapping-package}
\title{gsm.mapping: Good Statistical Monitoring Mapping}
\description{
The Good Statistical Monitoring or 'gsm' R package provides a framework for statistical data monitoring. 'gsm.mapping' provides workflows to apply the necessary data transformation from raw/source datasets to appropriate domains.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gilead-BioStats/gsm.mapping}
  \item \url{https://gilead-biostats.github.io/gsm.mapping}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gsm.mapping/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Jeremy Wildfire \email{jwildfire@gmail.com}
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
  \item Zelos Zhu \email{zelos.zhu@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
