% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_death.R
\name{complete_death}
\alias{complete_death}
\title{combines study completion data, overall response data, and death data to output a dataframe containing pd date and death status.}
\usage{
complete_death(
  dfDeath,
  dfStudyCompletion,
  dfOverallResponse,
  dfRandomization,
  chrDeathDateCol = "death_dt",
  chrStudyDiscontinuationReasonCol = "compreas",
  chrDeathResponse = "Death",
  chrResponseCol = "ovrlresp",
  chrPDResponse = "PD",
  chrResponseDateCol = "rs_dt",
  chrRandomizationDateCol = "rgmn_dt"
)
}
\arguments{
\item{dfDeath}{\code{data.frame} death data frame with mapped names}

\item{dfStudyCompletion}{\code{data.frame} study completion data frame with mapped names}

\item{dfOverallResponse}{\code{data.frame} overall response data frame with mapped names}

\item{dfRandomization}{\code{data.frame} randomization data frame with mapped names}

\item{chrDeathDateCol}{\code{character} name of column in \code{dfDeath} that contains
the death date. Default: \code{"death_dt"}.}

\item{chrStudyDiscontinuationReasonCol}{\code{character} name of column in \code{dfStudyCompletion} that contains
the reason for study completion. Default: \code{"compreas"}.}

\item{chrDeathResponse}{\code{character} value of \code{compreas} in \code{dfStudyCompletion} that indicates death. Default is "Death"}

\item{chrResponseCol}{\code{character} name of column in \code{dfOverallResponse} that contains
the overall response. Default: \code{"ovrlresp"}.}

\item{chrPDResponse}{\code{character} value of \code{ovrlresp} in \code{dfOverallResponse} that indicates PD. Default is "PD"}

\item{chrResponseDateCol}{\code{character} name of column in \code{dfOverallResponse} that contains
the overall response date. Default: \code{"rs_dt"}.}

\item{chrRandomizationDateCol}{\code{character} name of column in \code{dfRandomization} that contains the randomization date.
Default: \code{"rgmn_dt"}.}
}
\value{
a \code{data.frame} combining death status and first PD status
}
\description{
combines study completion data, overall response data, and death data to output a dataframe containing pd date and death status.
}
\examples{
\dontrun{
lSource <- list(
  Source_OverallResponse = gsm.endpoints::lSource_ep$Raw_OverallResponse, ## Overall response
  Source_Death = gsm.endpoints::lSource_ep$Raw_Death, ## Death date
  Source_STUDCOMP = gsm.endpoints::lSource_ep$Raw_STUDCOMP
)

lMapping <- gsm.core::MakeWorkflowList(
  strPath = "workflow/1_mappings",
  strNames = c("Death", "OverallResponse", "STUDCOMP"),
  strPackage = "gsm.mapping"
)

combined_spec <- gsm.mapping::CombineSpecs(lMapping)

lRaw <- gsm.mapping::Ingest(lSourceData = lSource, lSpec = combined_spec)

complete_death(
  dfDeath = lRaw$Raw_Death,
  dfStudyCompletion = lRaw$Raw_STUDCOMP,
  dfOverallResponse = lRaw$Raw_OverallResponse
)
}

}
