% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ingest.R
\name{Ingest}
\alias{Ingest}
\title{Ingests data from a source based on a given specification.}
\usage{
Ingest(lSourceData, lSpec, strDomain = "Raw")
}
\arguments{
\item{lSourceData}{\code{list} A named list of source data frames.}

\item{lSpec}{\code{list} A named list of column specifications.}

\item{strDomain}{\code{character} Domain name to add to the data frames after ingestions. Default: "Raw"}
}
\value{
\code{list} A named list of data frames, where each data frame corresponds to a domain in the
specification.
}
\description{
This function ingests data from a source based on a given specification. It iterates over each
domain in the specification and checks if the columns exist in the source data. It then
constructs a query to select the columns from the source and calls the \code{RunQuery} function to
retrieve the data. The resulting data frames are stored in a list, where each data frame
corresponds to a domain in the specification.
}
\examples{
core_mappings <- c("AE", "COUNTRY", "DATACHG", "DATAENT", "ENROLL", "LB",
                   "PD", "PK", "QUERY", "STUDY", "STUDCOMP", "SDRGCOMP", "SITE", "SUBJ")

basic_sim <- gsm.datasim::generate_rawdata_for_single_study(
  SnapshotCount = 1,
  SnapshotWidth = "months",
  ParticipantCount = 30,
  SiteCount = 5,
  StudyID = "ABC",
  workflow_path = "workflow/1_mappings",
  mappings = core_mappings,
  package = "gsm.mapping",
  desired_specs = NULL
)

lSourceData <- basic_sim[[1]]

lIngestWorkflow <- gsm::MakeWorkflowList(strName = core_mappings, strPath = "workflow/1_mappings", strPackage = "gsm.mapping")[[1]]
lRawData <- Ingest(lSourceData, lIngestWorkflow$spec)

}
